Move `EvaluationContext` to `org.gradle.internal.evaluation` package
