+++ platforms/core-configuration/model-core/src/main/java/org/gradle/api/internal/provider/FlatMapProvider.java
@@ -18,6 +18,7 @@ package org.gradle.api.internal.provider;
 
 import org.gradle.api.Transformer;
 import org.gradle.api.provider.Provider;
+import org.gradle.internal.evaluation.EvaluationContext;
 
 import javax.annotation.Nullable;
