+++ platforms/core-configuration/model-core/src/main/java/org/gradle/api/internal/provider/OrElseProvider.java
@@ -16,6 +16,8 @@
 
 package org.gradle.api.internal.provider;
 
+import org.gradle.internal.evaluation.EvaluationContext;
+
 import javax.annotation.Nullable;
 
 class OrElseProvider<T> extends AbstractMinimalProvider<T> {
